image-processing/lambda/requirements.txt
requests
pillow==10.2.0
boto3==1.34.23